refactor: consolidate test files and remove unit tests covered by integration
Remove all unit tests from shell.rs (all behaviors covered at integration
level via subprocess harness) and all executor unit tests except the IO
thread panic test (unreachable from subprocess).

Consolidate test files: edge_cases.rs absorbed into repl.rs; run_script.rs
and exit_validation.rs replaced by exit.rs which covers all exit behavior
including the "exit stops subsequent commands" scenario.

Remove exit_nonzero_code_propagates from repl.rs (redundant with
exit_valid_code_propagates in exit.rs).
