+++ src/executor.rs
@@ -490,205 +490,6 @@ fn join_io_thread(
 #[cfg(test)]
 mod tests {
     use super::*;
-    use std::path::PathBuf;
-
-    fn test_ctx() -> ShellCtx {
-        ShellCtx::new(None, std::env::current_dir().unwrap_or_else(|_| PathBuf::from("/")))
-    }
-
-    fn exec_builtin(builtin: BuiltInName, args: Vec<Arg>) -> ShellResult<Option<ExitCode>> {
-        let mut out: Vec<u8> = Vec::new();
-        let mut err: Vec<u8> = Vec::new();
-        let mut ctx = test_ctx();
-        execute(Command::BuiltIn(BuiltInCommand::new(builtin)), args, &mut out, &mut err, &mut ctx, None, None)
-    }
-
-    // --- exit ---
-
-    #[test]
-    fn exit_no_args_returns_success_exit_code() {
-        assert_eq!(exec_builtin(BuiltInName::Exit, vec![]).unwrap(), Some(ExitCode::SUCCESS));
-    }
-
-    #[test]
-    fn exit_with_numeric_arg_returns_that_code() {
-        assert_eq!(exec_builtin(BuiltInName::Exit, vec![Arg::from("42")]).unwrap(), Some(ExitCode(42)));
-    }
-
-    #[test]
-    fn exit_with_invalid_arg_writes_error_to_stderr() {
-        let mut out: Vec<u8> = Vec::new();
-        let mut err: Vec<u8> = Vec::new();
-        let mut ctx = test_ctx();
-        let result = execute(
-            Command::BuiltIn(BuiltInCommand::new(BuiltInName::Exit)),
-            vec![Arg::from("notanumber")],
-            &mut out, &mut err, &mut ctx, None, None,
-        );
-        assert!(result.is_ok());
-        assert_eq!(result.unwrap(), Some(ExitCode::FAILURE));
-        assert!(String::from_utf8_lossy(&err).contains("numeric argument required"));
-        assert!(out.is_empty(), "nothing written to stdout on bad exit arg");
-    }
-
-    // --- echo ---
-
-    #[test]
-    fn echo_no_args_writes_newline() {
-        let mut out: Vec<u8> = Vec::new();
-        execute_echo(vec![], &mut out).unwrap();
-        assert_eq!(out, b"\n");
-    }
-
-    #[test]
-    fn echo_single_arg_writes_arg_and_newline() {
-        let mut out: Vec<u8> = Vec::new();
-        execute_echo(vec![Arg::from("hello")], &mut out).unwrap();
-        assert_eq!(out, b"hello\n");
-    }
-
-    #[test]
-    fn echo_multiple_args_space_joins_them() {
-        let mut out: Vec<u8> = Vec::new();
-        execute_echo(vec![Arg::from("a"), Arg::from("b"), Arg::from("c")], &mut out).unwrap();
-        assert_eq!(out, b"a b c\n");
-    }
-
-    // --- pwd ---
-
-    #[test]
-    fn pwd_writes_cwd_to_stdout() {
-        let cwd = PathBuf::from("/some/test/path");
-        let ctx = ShellCtx::new(None, cwd.clone());
-        let mut out: Vec<u8> = Vec::new();
-        execute_pwd(vec![], &mut out, &ctx).unwrap();
-        assert_eq!(out, format!("{}\n", cwd.display()).as_bytes());
-    }
-
-    // --- cd ---
-
-    #[test]
-    fn cd_no_args_changes_to_home() {
-        let td = tempfile::tempdir().expect("temp dir");
-        let home = td.path().to_path_buf();
-        let mut ctx = ShellCtx::new(Some(home.clone()), home.clone());
-        execute_cd(vec![], &mut ctx).unwrap();
-        // canonicalize both sides to handle symlinks (e.g. /var → /private/var on macOS)
-        let got = std::fs::canonicalize(&ctx.cwd).unwrap_or(ctx.cwd.clone());
-        let expected = std::fs::canonicalize(&home).unwrap_or(home.clone());
-        assert_eq!(got, expected);
-    }
-
-    #[test]
-    fn cd_missing_path_returns_file_not_found() {
-        let mut ctx = test_ctx();
-        let result = execute_cd(vec![Arg::from("/nonexistent_path_xyz_123")], &mut ctx);
-        assert!(matches!(result, Err(ShellError::FileNotFound { .. })));
-    }
-
-    #[test]
-    fn cd_to_file_returns_not_a_directory() {
-        let td = tempfile::tempdir().expect("temp dir");
-        let file = td.path().join("file.txt");
-        std::fs::write(&file, b"").expect("create file");
-        let mut ctx = ShellCtx::new(None, td.path().to_path_buf());
-        let result = execute_cd(vec![Arg::from(file.to_str().unwrap())], &mut ctx);
-        assert!(matches!(result, Err(ShellError::NotADirectory { .. })));
-    }
-
-    // --- type ---
-
-    #[test]
-    fn type_builtin_writes_correct_format() {
-        let mut out: Vec<u8> = Vec::new();
-        execute_type(vec![Arg::from("echo")], &mut out).unwrap();
-        assert_eq!(out, b"echo is a shell builtin\n");
-    }
-
-    #[test]
-    fn type_nonexistent_returns_command_not_found_error() {
-        let mut out: Vec<u8> = Vec::new();
-        let result = execute_type(vec![Arg::from("nonexistentcommand")], &mut out);
-        assert_eq!(result.unwrap_err(), ShellError::CommandNotFound { name: "nonexistentcommand".to_string() });
-    }
-
-    #[test]
-    fn type_no_args_returns_missing_operand() {
-        let mut out: Vec<u8> = Vec::new();
-        let result = execute_type(vec![], &mut out);
-        assert_eq!(result.unwrap_err(), ShellError::MissingOperand);
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn type_executable_on_path_writes_path() {
-        use std::fs;
-        use std::os::unix::fs::PermissionsExt;
-
-        let dir = tempfile::tempdir().expect("create temp dir");
-        let bin_path = dir.path().join("my_test_tool");
-        fs::write(&bin_path, b"#!/bin/sh\n").expect("write fake executable");
-        let mut perms = fs::metadata(&bin_path).unwrap().permissions();
-        perms.set_mode(0o755);
-        fs::set_permissions(&bin_path, perms).expect("set executable bit");
-
-        let original_path = std::env::var_os("PATH").unwrap_or_default();
-        let mut new_dirs: Vec<PathBuf> = vec![dir.path().to_path_buf()];
-        new_dirs.extend(std::env::split_paths(&original_path));
-        let new_path = std::env::join_paths(&new_dirs).expect("join paths");
-        // SAFETY: test-only, single-threaded context
-        unsafe { std::env::set_var("PATH", &new_path) };
-
-        let mut out: Vec<u8> = Vec::new();
-        let result = execute_type(vec![Arg::from("my_test_tool")], &mut out);
-
-        // SAFETY: test-only, single-threaded context
-        unsafe { std::env::set_var("PATH", &original_path) };
-
-        result.unwrap();
-        let output = String::from_utf8(out).unwrap();
-        assert!(output.contains("my_test_tool is"), "got: {output}");
-        assert!(output.contains(bin_path.to_str().unwrap()), "path missing from: {output}");
-    }
-
-    // --- redirect routing ---
-
-    #[test]
-    fn stdout_redirect_writes_to_file_not_terminal() {
-        let dir = tempfile::tempdir().expect("temp dir");
-        let target = dir.path().join("out.txt");
-        let mut ctx = ShellCtx::new(None, dir.path().to_path_buf());
-        let mut out: Vec<u8> = Vec::new();
-        let mut err: Vec<u8> = Vec::new();
-        let redir = Some(StdoutRedirection::new(RedirectMode::Overwrite, target.clone()));
-        execute(
-            Command::BuiltIn(BuiltInCommand::new(BuiltInName::Echo)),
-            vec![Arg::from("hello")],
-            &mut out, &mut err, &mut ctx, redir, None,
-        ).unwrap();
-        assert!(out.is_empty(), "terminal stdout should be empty when redirected");
-        assert_eq!(std::fs::read_to_string(&target).unwrap(), "hello\n");
-    }
-
-    #[test]
-    fn stderr_redirect_writes_to_file_not_terminal() {
-        let dir = tempfile::tempdir().expect("temp dir");
-        let target = dir.path().join("err.txt");
-        let mut ctx = ShellCtx::new(None, dir.path().to_path_buf());
-        let mut out: Vec<u8> = Vec::new();
-        let mut err: Vec<u8> = Vec::new();
-        let redir = Some(StderrRedirection::new(RedirectMode::Overwrite, target.clone()));
-        execute(
-            Command::BuiltIn(BuiltInCommand::new(BuiltInName::Exit)),
-            vec![Arg::from("notanumber")],
-            &mut out, &mut err, &mut ctx, None, redir,
-        ).unwrap();
-        assert!(err.is_empty(), "terminal stderr should be empty when redirected");
-        let contents = std::fs::read_to_string(&target).unwrap();
-        assert!(contents.contains("numeric argument required"), "got: {contents}");
-    }
-
-    // --- thread panic safety ---
 
     #[test]
     fn join_io_thread_panic_becomes_non_fatal_io_error() {
