+++ src/shell.rs
@@ -162,105 +162,3 @@ impl ShellBuilder {
     }
 }
 
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::io::Cursor;
-
-    #[test]
-    fn run_repl_executes_commands_and_captures_output() {
-        let input = b"echo hello world\n";
-        let mut reader = BufReader::new(Cursor::new(input.as_ref()));
-        let mut out: Vec<u8> = Vec::new();
-        let mut err: Vec<u8> = Vec::new();
-        Shell::builder().build().run_repl(&mut reader, &mut out, &mut err).unwrap();
-        assert!(String::from_utf8_lossy(&out).contains("hello world"));
-    }
-
-    #[test]
-    fn run_repl_with_custom_prompt_writes_prompt_to_stdout() {
-        let input = b"echo hi\n";
-        let mut reader = BufReader::new(Cursor::new(input.as_ref()));
-        let mut out: Vec<u8> = Vec::new();
-        let mut err: Vec<u8> = Vec::new();
-        Shell::builder()
-            .with_prompt("PROMPT> ".to_string())
-            .build()
-            .run_repl(&mut reader, &mut out, &mut err)
-            .unwrap();
-        assert!(String::from_utf8_lossy(&out).contains("PROMPT> "));
-    }
-
-    #[test]
-    fn run_repl_with_config_prompt_writes_prompt_to_stdout() {
-        let config = ShellConfig { prompt: ">> ".to_string(), ..Default::default() };
-        let input = b"echo hi\n";
-        let mut reader = BufReader::new(Cursor::new(input.as_ref()));
-        let mut out: Vec<u8> = Vec::new();
-        let mut err: Vec<u8> = Vec::new();
-        Shell::builder().with_config(config).build().run_repl(&mut reader, &mut out, &mut err).unwrap();
-        assert!(String::from_utf8_lossy(&out).contains(">> "));
-    }
-
-    #[test]
-    fn run_script_executes_lines_and_captures_output() {
-        let mut out: Vec<u8> = Vec::new();
-        let mut err: Vec<u8> = Vec::new();
-        let code = Shell::builder().build().run_script(&["echo hello"], &mut out, &mut err).unwrap();
-        assert_eq!(code, ExitCode::SUCCESS);
-        assert_eq!(out.trim_ascii(), b"hello");
-    }
-
-    #[test]
-    fn run_script_stops_at_exit_skipping_subsequent_lines() {
-        let mut out: Vec<u8> = Vec::new();
-        let mut err: Vec<u8> = Vec::new();
-        Shell::builder().build().run_script(&["echo a", "exit", "echo b"], &mut out, &mut err).unwrap();
-        let stdout = String::from_utf8_lossy(&out);
-        assert!(stdout.contains("a"), "expected 'a'");
-        assert!(!stdout.contains("b"), "expected 'b' not to run after exit");
-    }
-
-    #[test]
-    fn run_script_skips_empty_and_whitespace_lines() {
-        let mut out: Vec<u8> = Vec::new();
-        let mut err: Vec<u8> = Vec::new();
-        let code = Shell::builder().build().run_script(&["", "  ", "echo ok"], &mut out, &mut err).unwrap();
-        assert_eq!(code, ExitCode::SUCCESS);
-        assert!(String::from_utf8_lossy(&out).contains("ok"));
-    }
-
-    #[test]
-    fn run_script_propagates_exit_code() {
-        let mut out: Vec<u8> = Vec::new();
-        let mut err: Vec<u8> = Vec::new();
-        let code = Shell::builder().build().run_script(&["exit 42"], &mut out, &mut err).unwrap();
-        assert_eq!(code.0, 42);
-    }
-
-    #[test]
-    fn run_script_empty_slice_returns_success() {
-        let mut out: Vec<u8> = Vec::new();
-        let mut err: Vec<u8> = Vec::new();
-        let code = Shell::builder().build().run_script(&[], &mut out, &mut err).unwrap();
-        assert_eq!(code, ExitCode::SUCCESS);
-    }
-
-    #[test]
-    fn builder_with_prompt_after_with_config_last_call_wins() {
-        let config = ShellConfig { prompt: "config_prompt ".to_string(), ..Default::default() };
-        let input = b"echo hi\n";
-        let mut reader = BufReader::new(Cursor::new(input.as_ref()));
-        let mut out: Vec<u8> = Vec::new();
-        let mut err: Vec<u8> = Vec::new();
-        Shell::builder()
-            .with_config(config)
-            .with_prompt("final_prompt ".to_string())
-            .build()
-            .run_repl(&mut reader, &mut out, &mut err)
-            .unwrap();
-        let stdout = String::from_utf8_lossy(&out);
-        assert!(stdout.contains("final_prompt "), "got: {stdout}");
-        assert!(!stdout.contains("config_prompt"), "should be overridden, got: {stdout}");
-    }
-}
